+++ src/extension.ts
@@ -2,6 +2,5 @@ import * as vscode from 'vscode';
 import { TreeTodoEditorProvider } from './treeTodoEditor.js';
 
 export function activate(context: vscode.ExtensionContext) {
-	// Register our custom editor providers
-	context.subscriptions.push(TreeTodoEditorProvider.register(context));
+  context.subscriptions.push(TreeTodoEditorProvider.register(context));
 }